refactor: move csvConfig to export.ts and clean up export logic
