--- web/core/components/analytics/config.ts
@@ -1,9 +1,0 @@
-import { mkConfig } from "export-to-csv";
-
-export const csvConfig = (workspaceSlug: string) =>
-  mkConfig({
-    fieldSeparator: ",",
-    filename: `${workspaceSlug}-analytics`,
-    decimalSeparator: ".",
-    useKeysAsHeaders: true,
-  });
