+++ web/core/components/analytics/export.ts
@@ -1,13 +1,23 @@
 import { ColumnDef, Row } from "@tanstack/react-table";
-import { download, generateCsv } from "export-to-csv";
-import { csvConfig } from "./config";
+import { download, generateCsv, mkConfig } from "export-to-csv";
+
+export const csvConfig = (workspaceSlug: string) =>
+  mkConfig({
+    fieldSeparator: ",",
+    filename: `${workspaceSlug}-analytics`,
+    decimalSeparator: ".",
+    useKeysAsHeaders: true,
+  });
 
 export const exportCSV = <T>(rows: Row<T>[], columns: ColumnDef<T>[], workspaceSlug: string) => {
   const rowData = rows.map((row) => {
-    const exportColumns = columns.map((col) => col.meta?.export).filter((v) => v != undefined);
-    const cells = exportColumns.reduce((acc, col) => {
-      const cell = col?.value(row) ?? "-";
-      return { ...acc, [col?.label ?? col?.key]: cell };
+    const exportColumns = columns.map((col) => col.meta?.export);
+    const cells = exportColumns.reduce((acc: Record<string, string | number>, col) => {
+      if (col) {
+        const cell = col?.value(row) ?? "-";
+        acc[col.label ?? col.key] = cell;
+      }
+      return acc;
     }, {});
     return cells;
   });
